Construction & conversion

Kmers.Kmer — Type

Kmer{A<:Alphabet,K,N} <: BioSequence{A}

A parametric, immutable, bitstype for representing Kmers - short sequences. Given the number of Kmers generated from raw sequencing reads, avoiding repetetive memory allocation and triggering of garbage collection is important, as is the ability to effectively pack Kmers into arrays and similar collections.

In practice that means we an immutable bitstype as the internal representation of these sequences. Thankfully, this is not much of a limitation - kmers are rarely manipulated and so by and large don't have to be mutable.

Excepting their immutability, they fulfill the rest of the API and behaviours expected from a concrete BioSequence type, and non-mutating transformations of the type are still defined.

Warning

Given their immutability, setindex and mutating sequence transformations are not implemented for Kmers e.g. reverse_complement!.

Tip

Note that some sequence transformations that are not mutating are available, since they can return a new kmer value as a result e.g. reverse_complement.

From other sequences

Type stable

julia> seq = LongSequence("TTAGC")
5nt DNA Sequence:
TTAGC

julia> DNAKmer{5}(seq)
DNA 5-mer:
TTAGC

Not type stable

julia> seq = LongSequence("TTAGC")
5nt DNA Sequence:
TTAGC

julia> DNAKmer(seq)
DNA 5-mer:
TTAGC

Tip

The difference here between achieving type stability or not is in providing the length (if you know it) as a parameter

From strings

Kmers can be constructed from strings using their constructors:

Type stable

julia> DNAKmer{8}("ATCGATCG")
DNA 8-mer:
ATCGATCG

julia> RNAKmer{8}("AUCGAUCG")
RNA 8-mer:
AUCGAUCG

Not type stable

julia> DNAKmer("ATCGATCG")
DNA 8-mer:
ATCGATCG

julia> RNAKmer("AUCGAUCG")
RNA 8-mer:
AUCGAUCG

Tip

The difference here between achieving type stability or not is in providing the length (if you know it) as a parameter

From arrays of BioSymbols

Kmers can be constructed using vectors or arrays of a BioSymbol type:

Type stable

julia> DNAKmer{5}([DNA_T, DNA_T, DNA_A, DNA_G, DNA_C])
DNA 5-mer:
TTAGC

julia> RNAKmer{5}([RNA_U, RNA_U, RNA_A, RNA_G, RNA_C])
RNA 5-mer:
UUAGC

Not type stable

julia> DNAKmer([DNA_T, DNA_T, DNA_A, DNA_G, DNA_C])
DNA 5-mer:
TTAGC

julia> RNAKmer([RNA_U, RNA_U, RNA_A, RNA_G, RNA_C])
RNA 5-mer:
UUAGC

Tip

The difference here between achieving type stability or not is in providing the length (if you know it) as a parameter